Copy actual node counts from node pools into status
This patch changes the controller so that it copies the
`status.replicas` field of the `NodePool` object into the
`status.nodeCounts[].numberOfNodes` field of the `ClusterOrder` object.

Note that currently there is no way to know what is the `resourceClass`
that corresponds to a node pool, so the best we can do is assume that
there will be only one item in the `status.nodeRequests` field of the
`CluserOrder` object, and only one corresponding node pool.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>